defined coordinates based on a common x-value
Filtered the coordinates using a user-defined x-value. The result resembles the airfoil of the blade (but is not due to the angle between the blade and the center of the hub.

Moved the code that writes the coordinates to a file into a function. Also turned the code tht finds the coordinatesinto a function as well.